app/requirements.txt
Flask==2.2.0
psycopg2-binary==2.9.3
pandas==1.3.5